refactor: extract shared runMigrations/resolveDbPath helper
Collapse the duplicated foreign-key-toggling migration block from
container.ts, admin.ts, reset-database.ts and run-migrations.ts into the new
db/migrate.ts.

The hardcoded DATABASE_URL default also moves into the helper, so all four
callers now delegate path resolution to resolveDbPath(). reset-database.ts
previously kept its own raw URL for a production-safety check that matched
"prod" in the DB file name; that check was fragile, so it is dropped — the
NODE_ENV === "production" guard is the sole safeguard.

Also rename scripts/migrate.ts to scripts/run-migrations.ts (updating the
dev-migrate-up Makefile target) so the runnable CLI no longer shares a
basename with the db/migrate.ts helper.
